Add PGXN and GitHub release workflow
Add `CHANGELOG.md` and fill it out for the forthcoming v0.1.0.

Add `.ci/mknotes` to parse `CHANGELOG.md` and generate a change log only
for the current version in `target/release_notes.md` (ported from a
personal Go project).

Add `META.json.in` and a `make` target to generate `META.json` to
release on PGXN.

Add `github/workflows/release.yml` to release to PGXN and GitHub.

Add other `make` targets to build `target/release_notes.md` and a
release Zip file. Although the `pgxn-bundle` script in the workflow does
the same thing, this is useful for testing that the bundle is as it
should be.