Disable the file loader and add installcheck
Replace the boon file loader with a no-op loader. Add tests to be sure
it actually does its thing; without it added to the compiler, the test
returns a file not found error! Long-term it'd be nice to prevent it
from being compiled at all; see santhosh-tekuri/boon#12 for the feature
request.

While at it, update the `error!()` and `info!()` calls to include the
failure reason. The magic `:#` suffix in the format does it.

Spotted while adding regression tests, also included here and run
through PGXS's `installcheck` make target. Go ahead and start the
apt-get-installed cluster in the `lint-and-test.yml` workflow and use it
for `make installcheck`.