Upgrade to pgrx v0.11.4 and use `include_bytes!()`
The new version of pgrx allows tests to be run as a different user,
which will allow pgrx testing support to be added to
/pgxn/docker-pgxn-tools. The change here lets me actually test it before
release.

As part of that testing, with the help of @eeeebbbbrrrr, switch to
loading the test JSON schemas into the test module at compile time
instead of runtime. This fixes an issue where the `CARGO_MANIFEST_DIR`
environment variable was not present while running tests as a different
user. It's more efficient anyway.