Use pgxn-tools in the Lint and Test workflow
v1.6.0 of the pgxn/pgxn-tools Docker image includes the Rust toolchain
(with clippy and rustfmt) and a new command `pgrx-build-test` that
builds and tests pgrx extensions. It will also run `make installcheck`
if it detects it.

So use that image to run all the tests, and remove the
`.ci/apt-install-postgres` script. Meanwhile, teach `.ci/test-cover` how
to handle `PGUSER` and `PGDATA`.